libct: split mount_entry into separate file
mount_entry has become complex and requires further refactoring.
Extract it into a separate file to improve modularity and maintainability.

No functional change, just moving the code around.
