+++ libcontainer/rootfs_linux.go
@@ -9,10 +9,8 @@ import (
 	"runtime"
 	"strconv"
 	"strings"
-	"syscall"
 	"time"
 
-	"github.com/cyphar/filepath-securejoin/pathrs-lite/procfs"
 	"github.com/moby/sys/mountinfo"
 	"github.com/moby/sys/userns"
 	"github.com/mrunalp/fileutils"
@@ -40,51 +38,6 @@ type mountConfig struct {
 	cgroup2Path     string
 	rootlessCgroups bool
 	cgroupns        bool
-}
-
-// mountEntry contains mount data specific to a mount point.
-type mountEntry struct {
-	*configs.Mount
-	srcFile *mountSource
-	dstFile *os.File
-}
-
-// srcName is only meant for error messages, it returns a "friendly" name.
-func (m mountEntry) srcName() string {
-	if m.srcFile != nil {
-		return m.srcFile.file.Name()
-	}
-	return m.Source
-}
-
-func (m mountEntry) srcStat() (os.FileInfo, *syscall.Stat_t, error) {
-	var (
-		st  os.FileInfo
-		err error
-	)
-	if m.srcFile != nil {
-		st, err = m.srcFile.file.Stat()
-	} else {
-		st, err = os.Stat(m.Source)
-	}
-	if err != nil {
-		return nil, nil, err
-	}
-	return st, st.Sys().(*syscall.Stat_t), nil
-}
-
-func (m mountEntry) srcStatfs() (*unix.Statfs_t, error) {
-	var st unix.Statfs_t
-	if m.srcFile != nil {
-		if err := unix.Fstatfs(int(m.srcFile.file.Fd()), &st); err != nil {
-			return nil, os.NewSyscallError("fstatfs", err)
-		}
-	} else {
-		if err := unix.Statfs(m.Source, &st); err != nil {
-			return nil, &os.PathError{Op: "statfs", Path: m.Source, Err: err}
-		}
-	}
-	return &st, nil
 }
 
 // needsSetupDev returns true if /dev needs to be set up.
@@ -534,77 +487,6 @@ func statfsToMountFlags(st unix.Statfs_t) int {
 		flags |= unix.MS_STRICTATIME
 	}
 	return flags
-}
-
-func (m *mountEntry) createOpenMountpoint(root *os.File) (Err error) {
-	rootfs := root.Name()
-	unsafePath := pathrs.LexicallyStripRoot(rootfs, m.Destination)
-	dstFile, err := pathrs.OpenInRoot(root, unsafePath, unix.O_PATH)
-	defer func() {
-		if dstFile != nil && Err != nil {
-			_ = dstFile.Close()
-		}
-	}()
-	if err == nil && m.Device == "tmpfs" {
-		// If the original target exists, copy the mode for the tmpfs mount.
-		stat, err := dstFile.Stat()
-		if err != nil {
-			return fmt.Errorf("check tmpfs source mode: %w", err)
-		}
-		dt := fmt.Sprintf("mode=%04o", syscallMode(stat.Mode()))
-		if m.Data != "" {
-			dt = dt + "," + m.Data
-		}
-		m.Data = dt
-	}
-	if err != nil {
-		if !errors.Is(err, unix.ENOENT) {
-			return fmt.Errorf("lookup mountpoint target: %w", err)
-		}
-
-		// If the mountpoint doesn't already exist, we want to create a mountpoint
-		// that makes sense for the source. For file bind-mounts this is an empty
-		// file, for everything else it's a directory.
-		dstIsFile := false
-		if m.Device == "bind" {
-			fi, _, err := m.srcStat()
-			if err != nil {
-				// Error out if the source of a bind mount does not exist as we
-				// will be unable to bind anything to it.
-				return err
-			}
-			dstIsFile = !fi.IsDir()
-		}
-		if dstIsFile {
-			dstFile, err = pathrs.CreateInRoot(root, unsafePath, unix.O_CREAT|unix.O_EXCL|unix.O_NOFOLLOW, 0o644)
-		} else {
-			dstFile, err = pathrs.MkdirAllInRoot(root, unsafePath, 0o755)
-		}
-		if err != nil {
-			return fmt.Errorf("make mountpoint %q: %w", m.Destination, err)
-		}
-	}
-
-	dstFullPath, err := procfs.ProcSelfFdReadlink(dstFile)
-	if err != nil {
-		return fmt.Errorf("get mount destination real path: %w", err)
-	}
-	if !pathrs.IsLexicallyInRoot(rootfs, dstFullPath) {
-		return fmt.Errorf("mountpoint %q is outside of rootfs %q", dstFullPath, rootfs)
-	}
-	if relPath, err := filepath.Rel(rootfs, dstFullPath); err != nil {
-		return fmt.Errorf("get relative path of %q: %w", dstFullPath, err)
-	} else if relPath == "." {
-		return fmt.Errorf("mountpoint %q is on the top of rootfs %q", dstFullPath, rootfs)
-	}
-	// TODO: Make checkProcMount use dstFile directly to avoid the need to
-	// operate on paths here.
-	if err := checkProcMount(rootfs, dstFullPath, *m); err != nil {
-		return fmt.Errorf("check proc-safety of %s mount: %w", m.Destination, err)
-	}
-	// Update mountEntry.
-	m.dstFile = dstFile
-	return nil
 }
 
 func mountToRootfs(c *mountConfig, m mountEntry) error {
@@ -1444,100 +1326,4 @@ func maskPaths(rootFs string, paths []string, mountLabel string) error {
 	}
 
 	return nil
-}
-
-func reopenAfterMount(rootFd, f *os.File, flags int) (_ *os.File, Err error) {
-	fullPath, err := procfs.ProcSelfFdReadlink(f)
-	if err != nil {
-		return nil, fmt.Errorf("get full path: %w", err)
-	}
-	rootfs := rootFd.Name()
-	if !pathrs.IsLexicallyInRoot(rootfs, fullPath) {
-		return nil, fmt.Errorf("mountpoint %q is outside of rootfs %q", fullPath, rootfs)
-	}
-	unsafePath := pathrs.LexicallyStripRoot(rootfs, fullPath)
-	reopened, err := pathrs.OpenInRoot(rootFd, unsafePath, flags)
-	if err != nil {
-		return nil, fmt.Errorf("re-open mountpoint %q: %w", unsafePath, err)
-	}
-	defer func() {
-		if Err != nil {
-			_ = reopened.Close()
-		}
-	}()
-
-	// NOTE: The best we can do here is confirm that the new mountpoint handle
-	// matches the original target handle, but an attacker could've swapped a
-	// different path to replace it. In the worst case this could result in us
-	// applying later vfsmount flags onto the wrong mount.
-	//
-	// This is far from ideal, but the only way of doing this in a race-free
-	// way is to switch the new mount API (move_mount(2) does not require this
-	// re-opening step, and thus no such races are possible).
-	reopenedFullPath, err := procfs.ProcSelfFdReadlink(reopened)
-	if err != nil {
-		return nil, fmt.Errorf("check full path of re-opened mountpoint: %w", err)
-	}
-	if reopenedFullPath != fullPath {
-		return nil, fmt.Errorf("mountpoint %q was moved while re-opening", unsafePath)
-	}
-	return reopened, nil
-}
-
-// Do the mount operation followed by additional mounts required to take care
-// of propagation flags. This will always be scoped inside the container rootfs.
-func (m *mountEntry) mountPropagate(rootFd *os.File, mountLabel string) error {
-	var (
-		data  = label.FormatMountLabel(m.Data, mountLabel)
-		flags = m.Flags
-	)
-	// Delay mounting the filesystem read-only if we need to do further
-	// operations on it. We need to set up files in "/dev", and other tmpfs
-	// mounts may need to be chmod-ed after mounting. These mounts will be
-	// remounted ro later in finalizeRootfs(), if necessary.
-	if m.Device == "tmpfs" || pathrs.LexicallyCleanPath(m.Destination) == "/dev" {
-		flags &= ^unix.MS_RDONLY
-	}
-
-	if err := utils.WithProcfdFile(m.dstFile, func(dstFd string) error {
-		return mountViaFds(m.Source, m.srcFile, m.Destination, dstFd, m.Device, uintptr(flags), data)
-	}); err != nil {
-		return err
-	}
-
-	// We need to re-open the mountpoint after doing the mount, in order for us
-	// to operate on the new mount we just created. However, we cannot use
-	// pathrs.Reopen because we need to re-resolve from the parent directory to
-	// get a new handle to the top mount.
-	//
-	// TODO: Use move_mount(2) on newer kernels so that this is no longer
-	// necessary on modern systems.
-	newDstFile, err := reopenAfterMount(rootFd, m.dstFile, unix.O_PATH)
-	if err != nil {
-		return fmt.Errorf("reopen mountpoint after mount: %w", err)
-	}
-	_ = m.dstFile.Close()
-	m.dstFile = newDstFile
-
-	// Apply the propagation flags on the new mount.
-	if err := utils.WithProcfdFile(m.dstFile, func(dstFd string) error {
-		for _, pflag := range m.PropagationFlags {
-			if err := mountViaFds("", nil, m.Destination, dstFd, "", uintptr(pflag), ""); err != nil {
-				return err
-			}
-		}
-		return nil
-	}); err != nil {
-		return fmt.Errorf("change mount propagation through procfd: %w", err)
-	}
-	return nil
-}
-
-func setRecAttr(m mountEntry) error {
-	if m.RecAttr == nil {
-		return nil
-	}
-	return utils.WithProcfdFile(m.dstFile, func(procfd string) error {
-		return unix.MountSetattr(-1, procfd, unix.AT_RECURSIVE, m.RecAttr)
-	})
 }
